Move CoT faithfulness into glassboxllms/experiments/
- Relocate module to glassboxllms/experiments/cot_faithfulness/
  per the library's folder structure (experiments/ for pipelines)
- Move demo notebook to notebooks/cot_faithfulness_demo.ipynb
- Update imports to use glassboxllms package path
- Remove standalone cot_faithfulness/ folder
