--- cot_faithfulness/requirements.txt
@@ -1,18 +1,0 @@
-# Core dependencies
-openai>=1.0.0
-datasets>=2.0.0
-matplotlib>=3.5.0
-numpy>=1.21.0
-
-# Development dependencies (optional)
-pytest>=7.0.0
-black>=22.0.0
-flake8>=4.0.0
-mypy>=0.950
-
-
-
-
-
-
-
