+++ glassboxllms/experiments/cot_faithfulness/questions/aqua.json
