+++ glassboxllms/experiments/cot_faithfulness/questions/arc_challenge.json
