+++ glassboxllms/experiments/cot_faithfulness/questions/mmlu.json
